Extract discovery, search, topics, coin_payments services
Thin corresponding endpoints and people list handlers that still
reached into ctx.materializer for profiles/questions/articles.
