+++ chafan_core/app/api/api_v1/endpoints/search.py
@@ -1,19 +1,12 @@
 from typing import Any, List
 
-
 from fastapi import APIRouter, Depends, Request, Response
 
-from chafan_core.app import crud, models, schemas
+from chafan_core.app import schemas
 from chafan_core.app.api import deps
-from chafan_core.app.services import sites as sites_service
-from chafan_core.app.services import submissions as submissions_service
 from chafan_core.app.infra.request_context import RequestContext
-from chafan_core.app.responders.question import preview_of_question_as_search_hit
 from chafan_core.app.limiter import limiter
-from chafan_core.utils.base import filter_not_none
-
-import logging
-logger = logging.getLogger(__name__)
+from chafan_core.app.services import search as search_service
 
 router = APIRouter()
 
@@ -29,10 +22,7 @@ def search_users(
     ctx: RequestContext = Depends(deps.get_request_context_logged_in),
     q: str,
 ) -> Any:
-    if q == "":
-        return []
-    users = crud.user.search_by_handle_or_full_name(ctx.get_db(), fragment=q)
-    return [ctx.preview_of_user(u) for u in users]
+    return search_service.search_users(ctx, q)
 
 
 @router.get("/sites/", response_model=List[schemas.Site])
@@ -44,10 +34,7 @@ def search_sites(
     ctx: RequestContext = Depends(deps.get_request_context_logged_in),
     q: str,
 ) -> Any:
-    if q == "":
-        return []
-    sites = crud.site.search(ctx.get_db(), fragment=q)
-    return [sites_service.site_schema(ctx, s) for s in sites]
+    return search_service.search_sites(ctx, q)
 
 
 @router.get("/topics/", response_model=List[schemas.Topic])
@@ -59,11 +46,7 @@ def search_topics(
     ctx: RequestContext = Depends(deps.get_request_context_logged_in),
     q: str,
 ) -> Any:
-    if q == "":
-        return []
-    return crud.topic.get_ilike(
-        ctx.get_db(), fragment=q, column=models.Topic.name
-    )
+    return search_service.search_topics(ctx, q)
 
 
 @router.get("/questions/", response_model=List[schemas.QuestionPreviewForSearch])
@@ -76,13 +59,7 @@ def search_questions(
     # This API is very time consuming! Must check user logged in
     q: str,
 ) -> Any:
-    if q == "":
-        return []
-    questions = crud.question.search(ctx.get_db(), q=q)
-# TODO no search hit limit
-    return filter_not_none(
-        [preview_of_question_as_search_hit(q) for q in questions]
-    )
+    return search_service.search_questions(ctx, q)
 
 
 @router.get("/articles/", response_model=List[schemas.ArticlePreview])
@@ -94,12 +71,7 @@ def search_articles(
     ctx: RequestContext = Depends(deps.get_request_context_logged_in),
     q: str,
 ) -> Any:
-    if q == "":
-        return []
-    articles = crud.article.search(ctx.get_db(), q=q)
-    return filter_not_none(
-        [ctx.materializer.preview_of_article(a) for a in articles]
-    )
+    return search_service.search_articles(ctx, q)
 
 
 @router.get("/submissions/", response_model=List[schemas.Submission])
@@ -111,12 +83,7 @@ def search_submissions(
     ctx: RequestContext = Depends(deps.get_request_context_logged_in),
     q: str,
 ) -> Any:
-    if q == "":
-        return []
-    submissions = crud.submission.search(ctx.get_db(), q=q)
-    return filter_not_none(
-        [submissions_service.submission_schema(ctx, q) for q in submissions]
-    )
+    return search_service.search_submissions(ctx, q)
 
 
 @router.get("/answers/", response_model=List[schemas.AnswerPreview])
@@ -128,9 +95,4 @@ def search_answers(
     ctx: RequestContext = Depends(deps.get_request_context_logged_in),
     q: str,
 ) -> Any:
-    if q == "":
-        return []
-    answers = crud.answer.search(ctx.get_db(), q=q)
-    return filter_not_none(
-        [ctx.materializer.preview_of_answer(a) for a in answers]
-    )
+    return search_service.search_answers(ctx, q)
