+++ chafan_core/app/api/api_v1/endpoints/topics.py
@@ -3,10 +3,10 @@ from typing import Any, List
 from fastapi import APIRouter, Depends
 from sqlalchemy.orm import Session
 
-from chafan_core.app import crud, models, schemas
+from chafan_core.app import schemas
 from chafan_core.app.api import deps
 from chafan_core.app.infra.request_context import RequestContext
-from chafan_core.utils.base import HTTPException_, filter_not_none
+from chafan_core.app.services import topics as topics_service
 
 router = APIRouter()
 
@@ -20,13 +20,7 @@ def get_topic(
     """
     Get topic.
     """
-    topic = crud.topic.get_by_uuid(db, uuid=uuid)
-    if topic is None:
-        raise HTTPException_(
-            status_code=400,
-            detail="The topic doesn't exist in the system.",
-        )
-    return topic
+    return topics_service.get_topic(db, uuid)
 
 
 @router.post("/", response_model=schemas.Topic)
@@ -40,7 +34,7 @@ def create_topic(
     Create new topic with name or return existing one.
     TODO: avoid spamming
     """
-    return crud.topic.get_or_create(db, name=topic_in.name)
+    return topics_service.create_topic(db, name=topic_in.name)
 
 
 @router.get("/{uuid}/questions/", response_model=List[schemas.QuestionPreview])
@@ -51,21 +45,8 @@ def get_topic_questions(
     skip: int = 0,
     limit: int = 100,
 ) -> Any:
-    db = ctx.get_db()
-    topic = crud.topic.get_by_uuid(db, uuid=uuid)
-    if topic is None:
-        raise HTTPException_(
-            status_code=400,
-            detail="The topic doesn't exist in the system.",
-        )
-    # FIXME: expensive
-    questions: List[models.Question] = topic.questions[skip : (skip + limit)]
-    return filter_not_none(
-        [
-            ctx.materializer.preview_of_question(question)
-            for question in questions
-            if not question.is_hidden
-        ]
+    return topics_service.list_topic_questions(
+        ctx, uuid=uuid, skip=skip, limit=limit
     )
 
 
@@ -75,10 +56,4 @@ def get_sub_topics(
     db: Session = Depends(deps.get_read_db),
     uuid: str,
 ) -> Any:
-    topic = crud.topic.get_by_uuid(db, uuid=uuid)
-    if topic is None:
-        raise HTTPException_(
-            status_code=400,
-            detail="The topic doesn't exist in the system.",
-        )
-    return topic.child_topics
+    return topics_service.list_sub_topics(db, uuid)
