+++ chafan_core/app/services/__init__.py
@@ -7,7 +7,9 @@ from chafan_core.app.services import (
     articles,
     audit,
     channels,
+    coin_payments,
     comments,
+    discovery,
     drafts,
     feedbacks,
     feed,
@@ -21,8 +23,10 @@ from chafan_core.app.services import (
     questions,
     reports,
     rewards,
+    search,
     sites,
     submissions,
+    topics,
     webhooks,
 )
 
@@ -33,7 +37,9 @@ __all__ = [
     "articles",
     "audit",
     "channels",
+    "coin_payments",
     "comments",
+    "discovery",
     "drafts",
     "feedbacks",
     "feed",
@@ -47,7 +53,9 @@ __all__ = [
     "questions",
     "reports",
     "rewards",
+    "search",
     "sites",
     "submissions",
+    "topics",
     "webhooks",
 ]
